refactor(repo): split backup_import.go into proxy/security/settings files
